+++ src/subclient.ts
@@ -742,7 +742,8 @@ export class SubClient {
     side: types.Side,
     orderType: types.OrderType = types.OrderType.LIMIT,
     clientOrderId = 0,
-    tag: String = constants.DEFAULT_ORDER_TAG
+    tag: String = constants.DEFAULT_ORDER_TAG,
+    blockhash?: string
   ): Promise<TransactionSignature> {
     let tx = new Transaction();
     let marketIndex = this._subExchange.markets.getMarketIndex(market);
@@ -785,7 +786,14 @@ export class SubClient {
     tx.add(orderIx);
 
     let txId: TransactionSignature;
-    txId = await utils.processTransaction(this._parent.provider, tx);
+    txId = await utils.processTransaction(
+      this._parent.provider,
+      tx,
+      undefined,
+      undefined,
+      undefined,
+      blockhash
+    );
     this._openOrdersAccounts[marketIndex] = openOrdersPda;
     return txId;
   }
@@ -810,7 +818,8 @@ export class SubClient {
     side: types.Side,
     orderType: types.OrderType = types.OrderType.LIMIT,
     clientOrderId = 0,
-    tag: String = constants.DEFAULT_ORDER_TAG
+    tag: String = constants.DEFAULT_ORDER_TAG,
+    blockhash?: string
   ): Promise<TransactionSignature> {
     let tx = new Transaction();
     let market = Exchange.getPerpMarket(this._asset).address;
@@ -854,7 +863,14 @@ export class SubClient {
     tx.add(orderIx);
 
     let txId: TransactionSignature;
-    txId = await utils.processTransaction(this._parent.provider, tx);
+    txId = await utils.processTransaction(
+      this._parent.provider,
+      tx,
+      undefined,
+      undefined,
+      undefined,
+      blockhash
+    );
     this._openOrdersAccounts[marketIndex] = openOrdersPda;
     return txId;
   }
@@ -884,6 +900,42 @@ export class SubClient {
       this._parent.publicKey,
       this._marginAccountAddress,
       this._openOrdersAccounts[marketIndex]
+    );
+  }
+
+  public createPlaceOrderInstruction(
+    marketIndex: number,
+    price: number,
+    size: number,
+    side: types.Side,
+    orderType: types.OrderType = types.OrderType.LIMIT,
+    clientOrderId = 0,
+    tag: String = constants.DEFAULT_ORDER_TAG
+  ): TransactionInstruction {
+    if (marketIndex == constants.PERP_INDEX) {
+      return this.createPlacePerpOrderInstruction(
+        price,
+        size,
+        side,
+        orderType,
+        clientOrderId,
+        tag
+      );
+    }
+
+    return instructions.placeOrderV3Ix(
+      this.asset,
+      marketIndex,
+      price,
+      size,
+      side,
+      orderType,
+      clientOrderId,
+      tag,
+      this.marginAccountAddress,
+      this._parent.publicKey,
+      this._openOrdersAccounts[marketIndex],
+      this._parent.whitelistTradingFeesAddress
     );
   }